third_party/blink/web_tests/svg/animations/rapid-syncbases.html

<!DOCTYPE html>

<title>Rapid syncbases</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<svg viewBox="0 0 250 50" xmlns="http://www.w3.org/2000/svg">
  <g id="a" opacity="0">
    <animate attributeName="opacity" from="0" to="1" begin="0ms;last.end+22ms" dur="5ms" fill="freeze"
      id="first"/>
    <animate attributeName="opacity" from="1" to="0" begin="last.end" dur="5ms" fill="freeze"/>
    <rect x="0" y="0" width="50" height="50" fill="#00F"/>
  </g>
  <g id="b" opacity="0">
    <animate attributeName="opacity" from="1" to="1" begin="first.end+5ms" dur="5ms" fill="freeze"/>
    <animate attributeName="opacity" from="1" to="0" begin="last.end" dur="5ms" fill="freeze"/>
    <rect x="50" y="0" width="50" height="50" fill="#00F"/>
  </g>
  <g id="c" opacity="0">
    <animate attributeName="opacity" from="1" to="1" begin="first.end+10ms" dur="5ms" fill="freeze"/>
    <animate attributeName="opacity" from="1" to="0" begin="last.end" dur="5ms" fill="freeze"/>
    <rect x="100" y="0" width="50" height="50" fill="#00F"/>
  </g>
  <g id="d" opacity="0">
    <animate attributeName="opacity" from="1" to="1" begin="first.end+15ms" dur="5ms" fill="freeze"/>
    <animate attributeName="opacity" from="1" to="0" begin="last.end" dur="5ms" fill="freeze"/>
    <rect x="150" y="0" width="50" height="50" fill="#00F"/>
  </g>
  <g id="e" opacity="0">
    <animate attributeName="opacity" from="1" to="1" begin="first.end+20ms" dur="70ms" fill="freeze"
      id="last" onbegin="messure()"/>
    <animate attributeName="opacity" from="1" to="0" begin="last.end" dur="5ms" fill="freeze"/>
    <rect x="200" y="0" width="50" height="50" fill="#00F"/>
  </g>
</svg>
<script>
  let onend_clearing = null;
  let onend_filling = null;

  async_test(t => {
    let svg = document.querySelector("svg");
    let elements = [
      document.querySelector("#a"),
      document.querySelector("#b"),
      document.querySelector("#c"),
      document.querySelector("#d"),
      document.querySelector("#e")
    ];
    console.log(elements);
    let passed = false;
    let filling_steps = 0;
    const num_runs = 5;
    messure = t.step_func(() => {
      filling_steps++;
      if (filling_steps == num_runs) {
        t.done();
        return;
      }
      assert_true(!passed, "1-1");
      passed = true;
      setTimeout(t.step_func((t) => {
        assert_true(passed, "1-2");
        passed = false;
        // Check that there are no holes in the animation.
        let found_empty = false;
        elements.forEach((e) => {
          let opacity = window.getComputedStyle(e, null).opacity;
          if (opacity < 0.5)
            found_empty = true;
          assert_true((opacity > 0.5 && !found_empty) || (0.5 > opacity && found_empty), "1-3");
        });
      }), 0);
    });
  });
</script>
